+++ src/util/mod.rs
@@ -12,27 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use druid::PlatformError;
-
-mod common;
-mod controller;
-mod model;
-mod storage;
-mod transforms;
-mod util;
-mod view;
-
-use controller::delegate::Delegate;
-use model::app::AppState;
-use view::window;
-
-fn main() -> Result<(), PlatformError> {
-    let window = window::window();
-
-    let data = AppState::new(window.id);
-
-    druid::AppLauncher::with_window(window)
-        .delegate(Delegate)
-        .use_env_tracing()
-        .launch(data)
-}
+pub mod shapes;
